requirements.txt
paho-mqtt
python-dotenv
pydantic-settings
resend
